requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
pinecone>=7.0.0
sentence-transformers>=2.2.2
google-generativeai>=0.3.2
langchain-text-splitters>=0.0.1
numpy>=1.26.0
requests>=2.31.0
huggingface-hub>=0.34.0
Pillow>=10.0.0
reportlab>=4.0.0 